Log strings in the console instead of objects.
Chrome has a stupid bug/feature that causes objects in the console to be empty (empty like Thomas' soul, not {}) when viewed after navigating away from the page that logged the object. Hence we're going to just log strings instead so people don't get confused and work on this bug for 4 hours and find out later everything was for naught.